Installing pass-otp with pacman

+++ install/pass.sh
@@ -18,22 +18,24 @@ install() {
   elif command -v apt-get > /dev/null 2>&1; then
     sudo apt-get install -y -qq tree
   elif command -v pacman > /dev/null 2>&1; then
-    sudo pacman -S --noconfirm --needed pass
+    sudo pacman -S --noconfirm --needed pass pass-otp
   fi
 
-  rm -rf "${HOME}/password-store"
-  git clone --depth 1 "https://git.zx2c4.com/password-store" "${HOME}/password-store"
-  pushd "${HOME}/password-store"
-  WITH_BASHCOMP=no WITH_ZSHCOMP=no WITH_FISHCOMP=no PREFIX="${HOME}/opt" make install
-  popd
-  rm -rf "${HOME}/password-store"
+  if ! command -v pass > /dev/null 2>&1; then
+    rm -rf "${HOME}/password-store"
+    git clone --depth 1 "https://git.zx2c4.com/password-store" "${HOME}/password-store"
+    pushd "${HOME}/password-store"
+    WITH_BASHCOMP=no WITH_ZSHCOMP=no WITH_FISHCOMP=no PREFIX="${HOME}/opt" make install
+    popd
+    rm -rf "${HOME}/password-store"
 
-  rm -rf "${HOME}/pass-otp"
-  git clone --depth 1 "https://github.com/tadfisher/pass-otp" "${HOME}/pass-otp"
-  pushd "${HOME}/pass-otp"
-  PREFIX="${HOME}/opt" BASHCOMPDIR=${HOME}/opt/bash_completion.d make install
-  popd
-  rm -rf "${HOME}/pass-otp"
+    rm -rf "${HOME}/pass-otp"
+    git clone --depth 1 "https://github.com/tadfisher/pass-otp" "${HOME}/pass-otp"
+    pushd "${HOME}/pass-otp"
+    PREFIX="${HOME}/opt" BASHCOMPDIR=${HOME}/opt/bash_completion.d make install
+    popd
+    rm -rf "${HOME}/pass-otp"
+  fi
 
   if ! command -v pass > /dev/null 2>&1; then
     return
